Replace `std::uint64_t` with `std::size_t` in tests for consistency with container APIs.

+++ tests/daw_hash_set_test.cpp
@@ -13,17 +13,17 @@
 #include <cstdint>
 
 void test_001( ) {
-	std::uint64_t count = 1024ULL;
+	std::size_t count = 1024ULL;
 	daw::hash_set_t<std::uint64_t> adapt( count );
-	for( std::uint64_t n = 0; n < count; ++n ) {
+	for( std::size_t n = 0; n < count; ++n ) {
 		adapt.insert( n );
 	}
 }
 
 void test_003( ) {
-	std::uint64_t count = 1024ULL;
+	std::size_t count = 1024ULL;
 	daw::hash_set_t<std::uint64_t> adapt( count );
-	for( std::uint64_t n = 0; n < count; ++n ) {
+	for( std::size_t n = 0; n < count; ++n ) {
 		adapt.insert( n );
 	}
 	for( std::uint64_t n = 0; n < count; ++n ) {
